Split `near-network-primitives` types related to `NetworkConfig` into `config.rs`
