+++ chain/network-primitives/src/lib.rs
@@ -1,3 +1,3 @@
+pub(crate) mod config;
 mod network_protocol;
 pub mod types;
-pub mod utils;
